Add Git as a category
Added a Git tag as well as the category (so that individual posts
can be tagged as belonging to the new category. This new category is for
all posts related to the Git source code management system. This is likely
to be aggregated ot the vcs-pkg planet.

Added the new category to the sidebar, for easier navigation.

Signed-off-by: Manoj Srivastava <srivasta@debian.org>